Rename FavoriteButtonComponent to FavoriteButton
Component files should not have the Component suffix.
